\name{chm}
\alias{chm}
\docType{data}
\title{LiDAR-derived Canopy Height Model - (CHM)}
\description{
LiDAR-derived Canopy Height Model - (CHM)
}
\usage{data(chm)}
\format{
  The format is:'RasterLayer'
}
\details{
The LiDAR-derived-CHM provided as an example dataset is from a longleaf pine forest at Eglin AFB, USA. 
}
\source{
The CHM was generated using Lastools software. The LiDAR data collection was funded by a grant (11-2-1-11) from the Joint Fire Science Program: Data set for fuels, fire behavior, smoke, and fire effects model development and evaluation-the RxCADRE project. R. Ottmar, PI; multiple Co-Is."
}
\references{
USDA Forest Service. Rocky Mountain Research Station - RMRS - Moscow, Idaho, USA.
}
\examples{
data(chm)
## plot(chm)
}
\keyword{chm}
