% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ForestCAS}
\alias{ForestCAS}
\title{LiDAR individual tree canopy area}
\usage{
ForestCAS(chm,loc,maxcrown,exclusion)
}
\arguments{
\item{chm}{A LiDAR-derived Canopy Height Model (CHM)}

\item{loc}{A dataframe of the individual tree location and height (x,y and height)}

\item{maxcrown}{Maximum individual tree crown radius expected. Default 10.0 m}

\item{exclusion}{Exclusion of the pixels that has values below the percent of maximum tree height. Default is 0.3}
}
\value{
returns Individual tree canopy boundary [[1]] and list of canopy area [[2]]
}
\description{
Compute and export individual tree canopy area detected on the LiDAR-derived Canopy Height Model (CHM)
}
\examples{
\dontrun{

#' Importing the LiDAR-derived CHM file
data(chm) # or set a CHM. e.g. chm<-readGDAL("CHM_stand.asc")

#' Set the loc parameter
sCHM<-CHMsmoothing(chm, filter="mean", ws=5) # smoothing CHM
loc<-sTreesCHM(sCHM, fws=5,htd=8) # or import a tree list

#' Set the maxcrown parameter
maxcrown=10.0

#' Set the exclusion parameter
exclusion=0.3 # 30\%

#' Compute individual tree detection canopy area
canopy<-ForestCAS(chm,loc, maxcrown,exclusion)

#' Getting the individual tree detection canopy area boundary
boundaryTrees<-canopy[[1]]
#' Plotting the individual tree canopy boundary over the CHM
plot(chm) # plotting CHM
plot(boundaryTrees, add=T, border='red', bg='transparent', asp=1) # adding tree canopy boundary to the plot

#' Getting the individual tree detection canopy area list
canopyList<-canopy[[2]]
summary(canopyList)
plot(SpatialPoints(canopyList[,1:2]),col="black", add=T, pch="*") # adding tree location to the plot
}
}
\author{
Carlos Alberto Silva
}

