% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getVolume.r
\name{getVolume}
\alias{getVolume}
\title{LiDAR - getVolume 3D}
\usage{
getVolume(xyz, id, alpha, plotCAS)
}
\arguments{
\item{xyz}{A 3-column matrix or dataframe with the x, y and z coordinates of the 3D LiDAR point cloud.}

\item{id}{A vector id for the x, y, z 3-column matrix.}

\item{alpha}{A single value or vector of values for \eqn{\alpha}. Its range from 0 to 1.}

\item{plotCAS}{Logical, if TRUE (default) plot the \eqn{\alpha}-shape 3D.}
}
\value{
Return volume of the 3D \eqn{\alpha}-shape of the LiDAR point cloud in units of cubic meters.
}
\description{
This function calculates the volume of the 3D \eqn{\alpha}-shape of the LiDAR point cloud.
}
\examples{
\dontrun{

# Importing LAS file:
LASfile <- system.file("extdata", "LASexample1.las", package="rLiDAR")

# Reading LAS file
LAS<-readLAS(LASfile,short=TRUE)

# Setring the xyz coordinates and subsetting the data
xyz<-subset(LAS[,1:3],LAS[,3] >= 1.37)

# Finding clusters
clLAS<-kmeans(xyz, 32)

# set the id vector
id<-as.factor(clLAS$cluster)

# set the alpha
alpha<-0.25

# set the plotCAS parameter
plotCAS=TRUE

# get the volume
volume<-getVolume(xyz=xyz, id=id, alpha=alpha, plotCAS=plotCAS)

# Add other plot parameters
aspect3d(1,1,0.5)
axes3d(c("x-", "x-", "y-", "z"), col="gray") # axes
title3d(xlab = "X Coord", ylab = " Y Coord", zlab = "Height", col="red") # title

}
#
}
\author{
Carlos Alberto Silva. Uses code by Beatriz Pateiro-Lopez (\emph{alphashape3d} package,see \code{\link[alphashape3d]{volume_ashape3d}})
}
\references{
Lafarge, T.; Pateiro-Lopez, B.; Possolo, A. and Dunkers, J.P. (2014). R Implementation of a Polyhedral Approximation to a 3D Set of Points Using the alpha-Shape Journal of Statistical Software Vol. 56(4), pp. 1-18.
}

