% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LiDARtrees3D}
\alias{LiDARtrees3D}
\title{LiDAR 3D trees stand visualization}
\usage{
LiDARtrees3D(crownshape = c("cone", "ellipsoid", "halfellipsoid",
                "paraboloid", "cylinder"), CL = 4, CW = 8, HCB = 10,
                  X = 0, Y = 0, dbh = 0.3, crowncolor = "forestgreen",
                 stemcolor = "chocolate4", resolution="high", shape=TRUE)
}
\arguments{
\item{crownshape}{Tree crown shape: "cone", "ellipsoid","halfellipsoid", "paraboloid" or "cylinder". Default is "halfellipsoid".}

\item{CL}{Tree crown length. Defaut is 4m.}

\item{CW}{Tree crown diameter. Defaut is 8 m.}

\item{HCB}{Tree trunk height. Defaut is 10 m.}

\item{X}{X Tree location. Defaut is 0.}

\item{Y}{Y Tree location. Defaut is 0.}

\item{dbh}{Tree stem base diameter. Defaut is 0.3 m.}

\item{crowncolor}{Tree crown color. Defaut is "forestgreen".}

\item{stemcolor}{Tree stem color. Defaut is "chocolate4".}

\item{resolution}{Tree crown resolution: "low", "median" and "high". Defaut is "high".}

\item{shape}{Logical, if TRUE (default) returns a tree crown mesh model, and if FALSE returns a tree crown line mode.}
}
\value{
returns 3D plot with single LiDAR trees
}
\description{
Draws a 3D scatterplot for LiDAR trees.
}
\examples{
\dontrun{

#---------------------------------------------------------#
# EXAMPLE 01: Plotting single trees

# shape cone crown
open3d()
LiDARtrees3D(crownshape = "cone", CL = 10, CW =7,
           HCB = 5, X =0, Y = 0, dbh = 0.4, crowncolor = "forestgreen",
                       stemcolor = "chocolate4", resolution="high", shape=T)

# elliptois crown shape
open3d()
LiDARtrees3D(crownshape = "ellipsoid", CL = 10, CW =7,
           HCB = 5, X =0, Y = 0, dbh = 0.4, crowncolor = "forestgreen",
                       stemcolor = "chocolate4", resolution="high", shape=T)

# halfellipsoid crown shape
open3d()
LiDARtrees3D(crownshape = "halfellipsoid", CL = 10, CW =7,
           HCB = 5, X =0, Y = 0, dbh = 0.4, crowncolor = "forestgreen",
                       stemcolor = "chocolate4", resolution="high", shape=T)

# paraboloid crown shape
open3d()
LiDARtrees3D(crownshape = "paraboloid", CL = 10, CW =7,
           HCB = 5, X =0, Y = 0, dbh = 0.4, crowncolor = "forestgreen",
                       stemcolor = "chocolate4", resolution="high", shape=T)

# Set the shape=FALSE
open3d()
LiDARtrees3D(crownshape = "paraboloid", CL = 10, CW =7,
           HCB = 5, X =0, Y = 0, dbh = 0.4, crowncolor = "forestgreen",
                       stemcolor = "chocolate4", resolution="high", shape=FALSE)

#---------------------------------------------------------#
#EXAMPLE 02: Plotting a virtual forest plantation stand

# Set the lenght of the stand
xlenght<-30 # x lenght
ylenght<-20 # y lenght

# Set the space between trees
sx<-3 # x space lengh
sy<-2 # y space lenght

# Tree location grid
XYgrid <- expand.grid(x = seq(1,xlenght,sx), y = seq(1,ylenght,sy))

# Get the number of trees
Ntrees<-nrow(XYgrid)

# Plotting a virtual Eucalyptus forest plantation stand using halfellipsoid tree crown shape

# Set stand trees parameters
meanHCB<-5 # mean tree crown base heigh
sdHCB<-0.1 # standard deviation tree crown base heigh
HCB<-rnorm(Ntrees, mean=meanHCB, sd=sdHCB) # tree crown base height
CL<-HCB # tree crown heigh
CW<-HCB*0.6 # tree crown diameter

open3d() # open a rgl window

# Plot stand
for( i in 1:Ntrees){
 LiDARtrees3D(crownshape = "halfellipsoid", CL = CL[i], CW = CW[i],
             HCB = HCB[i], X = XYgrid[i,1], Y = XYgrid[i,2], dbh = 0.4, crowncolor = "forestgreen",
                           stemcolor = "chocolate4", resolution="high", shape=T)
                           }

# Adding other plot parameters
axes3d(c("x-","x-", "y-","z"), col="gray") # axes
title3d(xlab = "X Coord", ylab = " Y Coord", zlab = "Height", col="red") # title
planes3d(a=0,b=0,c=-1,d=0.0001,color="gray",alpha=1) # set a terrain plane


# Plotting a virtual Pine forest plantation stand using cone tree crown shape

# Set stand trees parameters
meanHCB<-3 # mean tree crown base heigh
sdHCB<-0.1 # standard deviation tree crown base heigh
HCB<-rnorm(Ntrees, mean=meanHCB, sd=sdHCB) # tree crown base height
CL<-HCB*2.0 # tree crown heigh
CW<-HCB*1.3 # tree crown diameter

open3d() # open a rgl window
# Plot stand
for( i in 1:Ntrees){
 LiDARtrees3D(crownshape = "cone", CL = CL[i], CW = CW[i],
             HCB = HCB[i], X = XYgrid[i,1], Y = XYgrid[i,2], dbh = 0.4, crowncolor = "forestgreen",
                           stemcolor = "chocolate4", resolution="high", shape=T)
                           }

# Adding other plot parameters
axes3d(c("x-","x-", "y-","z"), col="gray") # axes
title3d(xlab = "X Coord", ylab = " Y Coord", zlab = "Height", col="red") # title
planes3d(a=0,b=0,c=-1,d=0.0001,color="gray",alpha=1) # set a terrain plane

#---------------------------------------------------------#
# EXAMPLE 03: Plotting a virtual natural mixed forest stand

# 01. Difers trees species on the stand using diferents crown shapes

# Set the number of trees
Ntrees<-80

# Set the trees locations
xcoord<-sample(1:100,Ntrees) # x coord
coord<-sample(1:100,Ntrees) # x coord

# Set a location grid of trees
XYgrid<-cbind(xcoord,ycoord)

# plot the location of the trees
plot(XYgrid, main="Tree location")

meanHCB<-7 # mean tree crown base heigh
sdHCB<-3 # standard deviation tree crown base heigh
HCB<-rnorm(Ntrees, mean=meanHCB, sd=sdHCB) # tree crown base heigh
crownshape<-sample(c("cone", "ellipsoid","halfellipsoid", "paraboloid"),
						Ntrees, replace=T) # tree crown shape
CL<-HCB*1.3 # tree crown heigh
CW<-HCB # tree crown diameter

open3d() # open a rgl window
# Plot stand

for( i in 1:Ntrees){
 LiDARtrees3D(crownshape = crownshape[i], CL = CL[i], CW = CW[i],
             HCB = HCB[i], X = XYgrid[i,1], Y = XYgrid[i,2], dbh = 0.4, crowncolor = "forestgreen",
                         stemcolor = "chocolate4", resolution="high", shape=T)
                         }

# Adding other plot parameters
axes3d(c("x-","x-", "y-","z"), col="gray") # axes
title3d(xlab = "X Coord", ylab = " Y Coord", zlab = "Height", col="red") # title
planes3d(a=0,b=0,c=-1,d=0.0001,color="gray",alpha=1) # set a terrain plane


# 02. Difers trees height on the stand using diferents crown colors

# Set the number of trees
Ntrees<-80

# Set the trees locations
xcoord<-sample(1:100,Ntrees) # x coord
ycoord<-sample(1:100,Ntrees) # x coord

# Set a location grid of trees
XYgrid<-cbind(xcoord,ycoord)

# plot the location of the trees
plot(XYgrid, main="Tree location")

meanHCB<-7 # mean tree crown base heigh
sdHCB<-3 # standard deviation tree crown base heigh
HCB<-rnorm(Ntrees, mean=meanHCB, sd=sdHCB) # tree crown base height
crownshape<-sample(c("cone", "ellipsoid","halfellipsoid", "paraboloid"),
                  Ntrees, replace=T) # tree crown shape
CL<-HCB*1.3 # tree crown heigh
CW<-HCB # tree crown diameter

# Plot tree height based on the HCB quantiles
HCBq<-quantile(HCB) # HCB quantiles
crowncolor<-NA*(1:Ntrees) # set a empty crowncolor vector

# classify trees by HCB quantile
for (i in 1:Ntrees){
 if (HCB[i] <= HCBq[2]) {crowncolor[i]<-"red"} # group 1
 if (HCB[i] > HCBq[2] & HCB[i] <= HCBq[3] ) {crowncolor[i]<-"blue"}  # group 2
 if (HCB[i] > HCBq[3] & HCB[i] <= HCBq[4] ) {crowncolor[i]<-"yellow"}  # group 3
 if (HCB[i] >= HCBq[4]) {crowncolor[i]<-"dark green"}  # group 4
 }

 open3d() # open a rgl window
# Plot stand
for( i in 1:Ntrees){
 LiDARtrees3D(crownshape = crownshape[i], CL = CL[i], CW = CW[i],
   HCB = HCB[i], X = XYgrid[i,1], Y = XYgrid[i,2], dbh = 0.4, crowncolor = crowncolor[i],
   stemcolor = "chocolate4", resolution="high", shape=T)
   }

# Adding other plot parameters
axes3d(c("x-","x-", "y-","z"), col="gray") # axes
title3d(xlab = "X Coord", ylab = " Y Coord", zlab = "Height", col="red") # title
planes3d(a=0,b=0,c=-1,d=0.0001,color="gray",alpha=1) # set a terrain plane
}
}
\author{
Carlos Alberto Silva and Remko Duursma. Uses code by Remko Duursma (MAESPA package,see \code{\link{MAESPA}}).
}
\references{
\code{\url{http://maespa.github.io/}}
}

