% Generated by roxygen2 (4.0.2): do not edit by hand
\name{singleTreeCHM}
\alias{singleTreeCHM}
\title{Individual tree detection on the LiDAR-derived Canopy Height Model (CHM)}
\usage{
singleTreeCHM(chm,fws,htd)
}
\arguments{
\item{chm}{A raster LiDAR-derived Canopy Height Model (CHM)}

\item{fws}{A single dimension of fixed square window size, e.g. 3, 5, 7 and so on. Default is 5.}

\item{htd}{Detection individual tree above specified heightbreak, e.g. 1.37, 2.0, 3.5 m and so on. Default is 1.37 m.}
}
\value{
returns A 3- column matrix of the individual tree dectection(x,y,and height)
}
\description{
Getting the individual tree location and hieght from the the LiDAR-derived Canopy Height Model (CHM)
}
\examples{
\dontrun{
# Importing the LiDAR-derived CHM file
data(chm) # or set a CHM. e.g. chm<-readGDAL("CHM_stand.asc")

# Smoothing CHM
schm<-CHMsmoothing(chm, "mean", 5, sigma=NULL)

# Plotting smothed CHM
plot(schm)

# Set the fws:
fws<-5 # dimention 5x5

# Set the specified heightbreak
htd<-8.0

# Getting the individual tree detection list
treeList<-singleTreeCHM(schm, fws, htd)
summary(treeList)

# Plotting the individual tree location on the CHM
plot(chm) # plotting CHM
plot(SpatialPoints(treeList[,1:2]), add=T, col="red") # plotthing tree location
}
}
\author{
Carlos Alberto Silva
}

