% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sTreesCHM}
\alias{sTreesCHM}
\title{Individual tree detection on the LiDAR-derived Canopy Height Model (CHM)}
\usage{
sTreesCHM(chm, fws,htd)
}
\arguments{
\item{chm}{A LiDAR-derived Canopy Height Model (CHM)}

\item{fws}{Single dimension of fixed square window size, e.g. 3, 5, 7 and so on. Default is 5.}

\item{htd}{Detection individual tree above specified heightbreak, e.g. 1.37, 2.0, 3.5 m and so on. Default is 1.37 m. m}
}
\value{
returns Individual tree dection list (x,y,and height)
}
\description{
Getting the individual tree location and hieght from the the LiDAR-derived Canopy Height Model (CHM)
}
\examples{
\dontrun{

#' Importing the LiDAR-derived CHM file
data(chm) # or set a CHM. e.g. chm<-readGDAL("CHM_stand.asc")

#' Set the fws:
fws<-5 # dimention 3x3

#' Set the specified heightbreak
htd<-8.0

#' Getting the individual tree detection list
treeList<-sTreesCHM(chm, fws,htd)
summary(treeList)

#' Plotting the individual tree location
plot(chm) # plotting CHM
plot(SpatialPoints(treeList[,1:2]), add=T, col="red") # plotthing tree location
}
}
\author{
Carlos Alberto Silva
}

