% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CHMsmoothing.r
\name{CHMsmoothing}
\alias{CHMsmoothing}
\title{LiDAR-derived Canopy Height Model (CHM) smoothing}
\usage{
CHMsmoothing(chm, filter, ws, sigma)
}
\arguments{
\item{chm}{A LiDAR-derived Canopy Height Model (CHM) raster file.}

\item{filter}{mean, median, maximum, minimum or gaussian. Default is "mean".}

\item{ws}{A single square matrix of weights dimension, e.g. 3,5, 7 and so on. Default is 5.}

\item{sigma}{Used only when filter parameter is equal to gaussian, e.g. 0.5, 1.0, 1.5 and so on. Default is 0.6.}
}
\value{
returns A CHM smoothed raster.
}
\description{
LiDAR-derived Canopy Height Model (CHM) smooth using focal statistcs
}
\examples{
\dontrun{

# Importing the LiDAR-derived CHM file
data(chm) # or set a CHM. e.g. chm<-raster("CHM_stand.asc")

#------------------------------------------------#
# Example 01: Smoothing CHM using gaussian filter

# Set the ws:
ws<-3 # dimention 3x3

# Set the filter type
filter<-"gaussian"

# Set the sigma value
sigma<-0.6

# Smoothing CHM
sCHM<-CHMsmoothing(chm, filter, ws, sigma)

# Plotting CHM smoothed
plot(sCHM, main=paste(filter,"filter and windows size", paste0(ws,"x",ws)))

#------------------------------------------------#
# Example 02: Smoothing CHM using mean filter

# Set the ws:
ws<-5 # dimention 5x5

# Set the filter type
filter<-"mean"

# Smoothing CHM
sCHM<-CHMsmoothing(chm, filter, ws, sigma=NULL)

# Plotting CHM smoothed
plot(sCHM, main=paste(filter,"filter and window size", paste0(ws,"x",ws)))
}
}
\author{
Carlos Alberto Silva.
}
\seealso{
\code{\link{focal}} in the \code{raster} package.
}

