% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestCAS.R
\name{forestCAS}
\alias{forestCAS}
\title{LiDAR-derived individual tree canopy area}
\usage{
forestCAS(chm,loc,maxcrown,exclusion)
}
\arguments{
\item{chm}{A LiDAR-derived Canopy Height Model (CHM) RasterLayer or SpatialGridDataFrame file.}

\item{loc}{A 3-column matrix or dataframe with the x,y coordinates and heights of the individual trees.}

\item{maxcrown}{A single value of the maximum individual tree crown radius expected. Default 10.0 m.}

\item{exclusion}{A single value from 0 to 1 that represent the \out{\%} of pixel exclusion. e.g. 0.5: It will exclude all the pixels for a single tree that has height value less 50\out{\%} of the maximum height from the same tree. Default is 0.3.}
}
\value{
returns A list that contains the individual tree canopy boundary polygons and the 3-column matrix with the x,y coordinates, heights and the canopy area (square meter).
}
\description{
Compute and export individual tree canopy area detected on the LiDAR-derived Canopy Height Model (CHM)
}
\examples{
\dontrun{

# Importing the LiDAR-derived CHM file
data(chm) # or set a CHM. e.g. chm<-raster("CHM_stand.asc")

# Set the loc parameter
sCHM<-CHMsmoothing(chm, filter="mean", ws=5) # smoothing CHM
loc<-singleTreeCHM(sCHM, fws=5,minht=8) # or import a tree list

# Set the maxcrown parameter
maxcrown=10.0

# Set the exclusion parameter
exclusion=0.3 # 30

# Compute individual tree detection canopy area
canopy<-forestCAS(chm,loc,maxcrown,exclusion)

#=======================================================================#
# Getting the individual tree detection canopy area boundary
#=======================================================================#
boundaryTrees<-canopy[[1]]
# Plotting the individual tree canopy boundary over the CHM
plot(chm) # plotting CHM
plot(boundaryTrees, add=T, border='red', bg='transparent') # adding tree canopy boundary

#=======================================================================#
# Getting the individual tree detection canopy area list
#=======================================================================#
canopyList<-canopy[[2]]
summary(canopyList)
plot(SpatialPoints(canopyList[,1:2]),col="black", add=T, pch="*") # adding tree location to the plot
}
}
\author{
Carlos Alberto Silva
}

