% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVolume}
\alias{getVolume}
\title{LiDAR - getVolume 3D}
\usage{
getVolume(xyz,id,alpha)
}
\arguments{
\item{xyz}{A 3-column matrix with the x,y and z coordinates of the LiDAR input points}

\item{alpha}{A value of the LiDAR input point id}

\item{cas}{Logical, if TRUE (default) plot the ??-shape.}

\item{alpha}{A single value or vector of values for ??.}
}
\value{
Return dataframe of the LAS data set
}
\description{
This function calculates the volume of the ??-shape of the LiDAR point cloud in the three-dimensional space using alphashape3d package.
}
\examples{
\dontrun{
#' Importing LAS file:
myLAS<-data(LASfile) # or set a LAS  file (myLAS<-"LASfile.las")

#' Reading LAS file
#'LAS<-readLAS(myLAS,short=TRUE)

#' Finding clusters
clLAS<-kmeans(LAS[,3], 32)

# set the xyz coordenates
xyz<-LAS[,1:3]

#' set the id vector
id<-as.factor(clLAS$cluster)

#' set the cas parameter
cas=TRUE

#' get the volume of the ??-shape
volume<-getVolume(xyz=xyz,id=id,cas=cas)

}
}
\author{
Carlos Alberto Silva. Uses code by Beatriz Pateiro-Lopez (alphashape3d package)
}
\references{
\link{http://cran.r-project.org/web/packages/alphashape3d/index.html}
}

