% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVolume}
\alias{getVolume}
\title{LiDAR - getVolume 3D}
\usage{
getVolume(xyz,id,alpha,cas)
}
\arguments{
\item{xyz}{A 3-column matrix with the x,y and z coordinates of the LiDAR input points}

\item{id}{A vector id for the the items in xyz.}

\item{alpha}{A single value or vector of values for alpha. Its range from 0 to 1.}

\item{cas}{Logical, if TRUE (default) plot the alpha-shape.}
}
\value{
Return dataframe of the LAS data set
}
\description{
This function calculates the volume of the alpha-shape of the LiDAR point cloud in the three-dimensional space using alphashape3d package.
}
\examples{
\dontrun{
# Importing LAS file:
myLAS<-data(LASfile) # or set a LAS  file (myLAS<-"LASfile.las")

# Reading LAS file
LAS<-readLAS(myLAS,short=TRUE)

# Finding clusters
clLAS<-kmeans(LAS[,3], 32)

# set the xyz coordenates
xyz<-LAS[,1:3]

# set the id vector
id<-as.factor(clLAS$cluster)

# set the alpha value
alpha<-0.25

# set the cas parameter
cas=TRUE

# get the volume of the alpha-shape
volume<-getVolume(xyz=xyz,id=id,alpha=alpha,cas=cas)
head(volume)
}
}
\author{
Carlos Alberto Silva. Uses code by Beatriz Pateiro-Lopez (alphashape3d R-package)
}
\references{
\link{http://cran.r-project.org/web/packages/alphashape3d/index.html}
}

