% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LASmetrics.r
\name{LASmetrics}
\alias{LASmetrics}
\title{LiDAR metrics computation}
\usage{
LASmetrics(LASfile,minht,above)
}
\arguments{
\item{LASfile}{A LAS standard data file}

\item{minht}{Use only returns above specified heightbreak, e.g. 1.30 m. Default is 1.37 m.}

\item{above}{Compute covers metrics using specified heightbreak, e.g. 2.5 m. Default is 2 m.}
}
\value{
Returns LiDAR-derived metrics acoording to McGaughey (2014)
}
\description{
Compute statistics metrics that describe the LiDAR data set
}
\examples{
\dontrun{

#---------------------------------------------------------#
# Example 01: Computing LiDAR metrics for a single LAS file

# Importing the LAS data file
LASfile<-paste0(path.package("rLiDAR"),"/data/LASexample.las") # In this case a height normalized LAS file

# Set the minht and above parameters
minht<-1.37 # meters or feet
above<-2.00  # meters or feet

# LiDAR metrics computation
LiDARmetrics<-LASmetrics(LASfile,minht,above)

#---------------------------------------------------------#
# Example 02: Computing LiDAR metrics for a multiple LAS files within a folder

# Set a folder where there are LAS files
folder="C:/lasfolder"

# Get a list of LAS file in the folder
LASlist <- list.files(folder, pattern="*.las", full.names=TRUE)

# Set the minht and above parameters
minht<-1.37 # meters or feet
above<-2.00  # meters or feet

# Creatin a empty dataframe to store the LiDAR metrics
LiDARmetrics<-data.frame()

# Set a loop to computate the LiDAR metrics
for ( i in LASlist) {
 LiDARmetrics<-rbind(LiDARmetrics,LASmetrics(i,minht,above))}
}
}
\author{
Carlos Alberto Silva
}
\seealso{
McGaughey, R. 2014. FUSION/LDV: Software for lidar data analysis and visualiZation. Version 3.41. Seattle, WA: U.S. Department of Agriculture, Forest Service, Pacific Northwest Research Station. Available at \code{\url{http://http://forsys.cfr.washington.edu/fusion/fusionlatest.html}}.
}

