% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readLAS}
\alias{readLAS}
\title{LiDAR - Read LiDAR data}
\usage{
readLAS(LASfile, short=TRUE)
}
\arguments{
\item{LASfile}{A standard LAS data file (ASPRS)}

\item{short}{Logical, if TRUE it will return only a 3-column matrix with the x, y, z, intensity and return number information}
}
\value{
Return a matrix of the information stored in the LAS file
}
\description{
Read LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{
\dontrun{
# Importing LAS file:
myLAS<-data(LASfile) # or set a LAS  file (myLAS<-"LASfile.las")

# Reading LAS file
readLAS(myLAS,short=TRUE)

# Summary of the LAS file
summary(myLAS)
}
}
\author{
Michael Sumner and Carlos Alberto Silva.
}
\references{
\code{\link{http://staff.acecrc.org.au/~mdsumner/las/readLAS.R}}
}

