% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readLAS}
\alias{readLAS}
\title{LiDAR - Importing and reading LAS file}
\usage{
readLAS(LASfile)
}
\arguments{
\item{LASfile}{A standard LAS data file}

\item{short}{Return only : x, y, z, intensity and return number}
}
\value{
Return dataframe of the LAS data set
}
\description{
Importing and reading LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{
\dontrun{
#' Importing LAS file:
myLAS<-data(LASfile) # or set a LAS  file (myLAS<-"../LASfile.las)"

#' Reading LAS file
readLAS(myLAS)

#' Summary of the LAS file
summary(myLAS)
}
}
\author{
Michael Sumner
}

