% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readLAS.r
\name{readLAS}
\alias{readLAS}
\title{LiDAR - Read LiDAR data}
\usage{
readLAS(LASfile, short=TRUE)
}
\arguments{
\item{LASfile}{A standard LAS data file (ASPRS)}

\item{short}{Logical, if TRUE it will return only a 3-column matrix with the x, y, z, intensity and return number information}
}
\value{
Return a matrix of the information stored in the LAS file
}
\description{
Read LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{
\dontrun{
# Importing LAS file:
LASfile<-paste0(path.package("rLiDAR"),"/data/LASexample.las")

# Reading LAS file
rLAS<-readLAS(LASfile,short=TRUE)

# Summary of the LAS file
summary(rLAS)

# plot 2D
plot(rLAS[,1],rLAS[,2], xlab="X coord", ylab="Y coord")

# plot 3D
require(rgl)
plot3d(rLAS[,1:3], col="forestgreen",xlab="X coord", ylab="Y coord")
}
}
\author{
Michael Sumner and Carlos Alberto Silva.
}

