% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readLAS.r
\name{readLAS}
\alias{readLAS}
\title{LiDAR - Read LiDAR data}
\usage{
readLAS(LASfile, short=TRUE)
}
\arguments{
\item{LASfile}{A standard LAS data file (ASPRS)}

\item{short}{Logical, if TRUE it will return only a 3-column matrix with the x, y, z, intensity and return number information}
}
\value{
Return a matrix of the information stored in the LAS file
}
\description{
Read LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{
\dontrun{

#=======================================================================#
# Importing LAS file:
#=======================================================================#
LASfile <- system.file("extdata", "LASexample1.las", package="rLiDAR")

# Reading LAS file
rLAS<-readLAS(LASfile,short=TRUE)

# Summary of the LAS file
summary(rLAS)

#=======================================================================#
# LAS file visualization:
#=======================================================================#
#_______________________________________________________________________#
# Set a sigle color
#_______________________________________________________________________#
col<-"forestgreen"

# plot 2D
plot(rLAS[,1],rLAS[,2], col=col,xlab="UTM.Easting", ylab="UTM.Northing", main="LAS example file")

# plot 3D
require(rgl)
plot3d(rLAS[,1:3], col=col, axes=F,xlab="", ylab="", zlab="")
axes3d(c("x-", "y-", "z-"))
title3d(xlab = "UTM.Easting", ylab = "UTM.Northing",zlab = "Height(m)", col="red")
planes3d(0,0,-1,0.001,col="red",alpha=0.7)

#_______________________________________________________________________#
# Set a color by Height
#_______________________________________________________________________#
# color ramp
myColorRamp <- function(colors, values) {
v <- (values - min(values))/diff(range(values))
x <- colorRamp(colors)(v)
rgb(x[,1], x[,2], x[,3], maxColorValue = 255)
}

# Color by height
col <- myColorRamp(c("blue","green","yellow","red"),rLAS[,3])

# plot 2D
plot(rLAS[,1],rLAS[,2], col=col,xlab="UTM.Easting", ylab="UTM.Northing",main="LAS file")

# plot 3D
plot3d(rLAS[,1:3], col=col, axes=F,xlab="", ylab="", zlab="")
axes3d(c("x-", "y-", "z-"))
title3d(xlab = "UTM.Easting", ylab = "UTM.Northing",zlab = "Height(m)", col="red")
planes3d(0,0,-1,0.001,col="red",alpha=0.7)
#'}
}
\author{
Michael Sumner and Carlos Alberto Silva.
}

