% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rMetrics.r
\name{rMetrics}
\alias{rMetrics}
\title{Individual tree LiDAR metrics computation}
\usage{
rMetrics(xyziId)
}
\arguments{
\item{xyziId}{A 5-column matrix with the x, y, z coordinates, intensity and the tree id classification for the LiDAR point cloud.}
}
\value{
Returns A matrix of the LiDAR-based metrics for the individual tree detected.
}
\description{
This funtion computes LiDAR-based statistic metrics for individual trees detected from the 3D LiDAR point cloud.
}
\details{
List of the individual tree LiDAR metrics
==========================================================
Acronym and Metrics description
==========================================================
TotalReturns-> Total number of returns;
Etop-> UTM Easting coordinate of the tree top;
Ntop-> UTM Northing coordinate of the tree top;
Emin-> Minimum UTM Easting coordinate;
Nmin-> Minimum UTM Northing coordinate;
Emax-> Maximum UTM Easting coordinate;
Nmax-> Maxmium UTM Northing coordinate;
Ewidth-> Tree crown width 01;
Nwidth-> Tree crown width 02;
hmax-> Maximum Height;
hmean-> Mean height;
hsd-> Standard deviation of height;
hcv-> Coefficient of variation of height;
hmod-> Mode of height;
h5-> 5th percentile of height;
h10-> 10th percentile of height;
h20-> 20th percentile of height;
h25-> 25th percentile of height;
h30-> 30th percentile of height;
h40-> 40th percentile of height;
h50-> 50th percentile of height;
h60-> 60th percentile of height;
h70-> 70th percentile of height;
h75-> 75th percentile of height;
h80-> 80th percentile of height;
h90-> 90th percentile of height;
h95-> 95th percentile of height;
h99-> 99th percentile of height;
imax->  Maximum intensity;
imean-> Mean intensity;
isd-> Standard deviation of intensity;
icv-> Coefficient of variation of intensity;
imod-> Mode of intensity;
i5-> 5th percentile of intensity;
i10-> 10th percentile of intensity;
i20-> 20th percentile of intensity;
i25-> 25th percentile of intensity;
i30-> 30th percentile of intensity;
i40-> 40th percentile of intensity;
i50-> 50th percentile of intensity;
i60-> 60th percentile of intensity;
i70-> 70th percentile of intensity;
i75-> 75th percentile of intensity;
i80-> 80th percentile of intensity;
i90-> 90th percentile of intensity;
i95-> 95th percentile of intensity;
i99-> 99th percentile of intensity.
}
\examples{
\dontrun{

#=======================================================================#
# Individual tree detection using K-means cluster
#=======================================================================#
# Importing LAS file:
LASfile <- system.file("extdata", "LASexample1.las", package="rLiDAR")

# Reading LAS file
LAS<-readLAS(LASfile,short=TRUE)

# Setting the xyz coordinates and subsetting the data
xyzi<-subset(LAS[,1:4],LAS[,3] >= 1.37)

# Finding clusters (trees)
clLAS<-kmeans(xyzi[,1:2], 32)

# Set the tree id vector
Id<-as.factor(clLAS$cluster)

# Combining xyzi and tree id
xyziId<-cbind(xyzi,Id)

#=======================================================================#
#  Computing individual tree LiDAR metrics
#=======================================================================#

TreesMetrics<-rMetrics(xyziId)

}
}
\author{
Carlos Alberto Silva
}

