% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleTreeCHM.R
\name{singleTreeCHM}
\alias{singleTreeCHM}
\title{Individual tree detection on the LiDAR-derived Canopy Height Model (CHM)}
\usage{
singleTreeCHM(chm,fws,minht)
}
\arguments{
\item{chm}{A LiDAR-derived Canopy Height Model (CHM) raster  file}

\item{fws}{A single dimension of fixed square window size, e.g. 3, 5, 7 and so on. Default is 5}

\item{minht}{Detect individual tree above specified heightbreak, e.g. 1.37, 2.0, 3.5 m and so on. Default is 1.37 m}
}
\value{
returns A 3-column matrix of the location and height for the individual trees dectected
}
\description{
Detect and compute individual individual tree location and height on the LiDAR-derived Canopy Height Model (CHM)
}
\examples{
\dontrun{

# Importing the LiDAR-derived CHM raster file
library(raster)
data(chm) # or set a CHM. e.g. chm<-raster("CHM_stand.asc")

# Smoothing CHM
schm<-CHMsmoothing(chm, "mean", 5, sigma=NULL)

# Plotting smothed CHM
plot(schm)

# Set the fws:
fws<-5 # dimention 5x5

# Set the specified heightbreak
minht<-8.0

# Getting the individual tree detection list
treeList<-singleTreeCHM(schm, fws, minht)
summary(treeList)

# Plotting the individual tree location on the CHM
library(sp)
plot(chm) # plotting CHM
plot(SpatialPoints(treeList[,1:2]), add=T, col="red") # plotthing tree location
}
}
\author{
Carlos Alberto Silva
}

