% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleTreeCHM.R
\name{singleTreeCHM}
\alias{singleTreeCHM}
\title{Individual tree detection whitin the LiDAR-derived Canopy Height Model (CHM)}
\usage{
singleTreeCHM(chm,fws,minht)
}
\arguments{
\item{chm}{A LiDAR-derived Canopy Height Model (CHM) raster  file}

\item{fws}{A single dimension (in raster grid cell units) of fixed square window size, e.g. 3, 5, 7 and so on. Default is 5}

\item{minht}{Detect individual trees above specified height break, e.g. 1.37, 2.0, 3.5 m and so on. Default is 1.37 m.}
}
\value{
Returns a 3-column matrix of the tree number, location, and height for each individual trees detected.
}
\description{
Detects and computes the location and height of individual trees within the LiDAR-derived Canopy Height Model (CHM). The algorithm implemented in this function is local maximum with a fixed window size.
}
\examples{
\dontrun{

# Importing the LiDAR-derived CHM raster file

library(raster)
data(chm) # or set a CHM. e.g. chm<-raster("CHM_stand.asc")

# Smoothing CHM
schm<-CHMsmoothing(chm, "mean", 5)

# Plotting smothed CHM
plot(schm)

# Setting the fws:
fws<-5 # dimention 5x5

# Setting the specified height above ground for detectionbreak
minht<-8.0

# Getting the individual tree detection list
treeList<-singleTreeCHM(schm, fws, minht)
summary(treeList)

# Plotting the individual tree location on the CHM
plot(chm, main="LiDAR-derived CHM")
plot(SpatialPoints(treeList[,1:2]), add=T, col="red") # plotthing tree location
}
}
\author{
Carlos Alberto Silva
}

